refactor(sonar): contratos de excepción honestos y limpieza mecánica
- throws Exception -> StripeException en toda la cadena de facturación (API,
  controller, caso de uso e impl): el contrato ahora dice lo que de verdad
  puede fallar en vez de obligar a capturar cualquier cosa.
- ObjectStorageService: MinIO declara nueve checked distintas y por eso la
  firma era throws Exception. Se unifican en ObjectStorageException, con la
  original como causa. Quien sube una imagen no puede actuar distinto según
  cuál de las nueve sea.
- ImageMirrorService: throws Exception -> UnknownHostException / IOException +
  InterruptedException, que es lo que se lanza.
- PayPalGateway: bodyToMono(Map.class) devolvía el tipo crudo y obligaba a
  castear con @SuppressWarnings; ahora usa ParameterizedTypeReference.
- Mockito.x() -> imports estáticos, assertThat(m.get(k)).isEqualTo(v) ->
  containsEntry, isZero, hasToString, Math.clamp, StringBuilder.isEmpty,
  casts y throws sobrantes.
- Se retiran las dos @Transactional que sí eran redundantes (appendEvent y
  limitFor: sólo se llaman desde métodos de su clase que ya abren transacción).
  Las otras 113 las necesitan sus llamadores externos y se dejan.
- pom: se ignora java:S1135. Los 30 avisos eran la palabra española "todo" en
  comentarios en castellano, no tareas pendientes; revisados uno a uno.

+++ src/main/java/com/nexaplatform/dropshipping/application/usecase/impl/OrderUseCaseImpl.java
@@ -611,7 +611,7 @@ public class OrderUseCaseImpl implements OrderUseCase {
         String method = req.getPaymentMethod() == null ? WALLET : req.getPaymentMethod().toUpperCase();
         Order o = orderRepository.findById(created.getId()).orElseThrow();
         if (WALLET.equals(method)) {
-            long charge = (long) created.getTotalCents();
+            long charge = created.getTotalCents();
             String idemKey = idem != null ? idem : ("checkout-" + created.getId());
             walletUseCase.charge(userId, charge, created.getId(), idemKey, "Order " + created.getOrderNumber());
             o.setStatus(OrderStatus.PAID);
